[backend] bs_admin: add option to replace a project's signing key
It is sometimes useful to be able to sign with existing keys, perhaps
due to corporate policies (EG: for EFI Secure Boot).
It is not enough to substitute a project's _signkey and _pubkey on the
filesystem, the database has to be updated or the previous _sslcert
will be used, since that is never stored on the disk.
Add a new --update-project-signing-key option to bs_admin to facilitate
the operation for an administrator.

(cherry picked from commit 22af60efcfff5d8f534a15d6d22455d61d92563c)